src-python/requirements.txt
# 同人写作辅助系统 - Python sidecar 依赖
# Phase 1 基础依赖

# Web 框架
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0

# 数据序列化
pyyaml>=6.0,<7.0
pydantic>=2.0,<3.0
python-frontmatter>=1.0,<2.0

# 向量数据库（Phase 1 核心依赖）
chromadb>=0.5.0,<1.0.0

# LLM API 客户端（OpenAI 兼容接口）
openai>=1.0,<2.0

# Token 计数
tiktoken>=0.7.0,<1.0

# 文件锁（ops.jsonl 并发写入保护）
filelock>=3.0,<4.0

# 打包分发
pyinstaller>=6.0,<7.0
